requirements.txt
langchain
langchain-openai
beautifulsoup4
requests
chromadb
python-dotenv
tiktoken
streamlit
Flask 
Flask-Cors